Adding a new flag for cleaning up old resources (#90)
* Adding a new flag to topgrade called "cleanup"

If you are using applications like homebrew they may keep a history of
packages downloaded. Especially the bigger your number of installed
packages for homebrew grows the bigger those directories get. This can
quickly add up to a couple of GBs.

If this flag is set then the homebrew part also calls "brew cleanup" for
cleaning up old resources. Of course this can be add to other calls as
well if supported.

* Updating readme for new cleanup flag