git/windows: fix output corruption when running `git submodule`
The `git submodule update` command, run as part of updating a git repo,
caused output corruption on Windows.  The effect was to disable
ANSI/VT100 control sequence processing on the terminal.

While the root cause isn't clear this is almost certainly caused by some
component of the mingw chain setting the Windows "console mode" field to
disable escape sequence processing as part of resetting it to "normal"
for the platform.

With some experimentation I determined that setting the stdin handle for
the `git submodule` process would alter the behaviour, preventing the
problem.  Again, no root cause, but this almost certainly changes the
understanding that some subprocess has to "no terminal" rather than
"attached to a terminal."

Only `Stdin::null()` worked for this; piped or inherited stdin for the
command would observe the same behaviour.

The fact it is a change to the console settings via a Windows APi of
some sort is clear because it applies regardless of any display of
output from the subcommand.  It can't be a control sequence, but can be
a direct modification of the Windows "pty" equivalent.

This fixes #538, reasonably robustly, by applying this to *all* git
command invocations; it could be applied only to submodule, but that
would leave open the possibility that another command is changed
internally and gained the same display-breaking behaviour.

It is possible that, in the future, whatever component triggers this may
start testing the `stdout` or `stderr` handles instead of `stdin` to
determine what to do.  If so, that'll show up as the problem
reoccurring, in which case the only reasonable fix I can see would be to
direct all input and output to files on disk.  -_-  Thanks, Win32.

Tested as thoroughly as I could, but only manually.  I can't figure out
a usefully robust automated test that'd work on Windows, let alone
portably.