Fix Microsoft Store (fix #656)
Despite the fact that Topgrade explicitly tries to run the Windows Powershell, gsudo has an odd behaviour which causes it to detect whether it's being run from Powershell for Cmd.
If run from Powershell, It will launch the current powershell with the given command. If the current shell is Powershell, it will try to run the command with it.
However, Get-WmiObject doesn't exist on Powershell core.