Allow to run `flatpak update` with sudo
This change adds the option `flatpak.use_sudo` that allows to update
the system-wide installation with sudo. When set to `true` the
system-wide installation will be updated with sudo. If set to `false`
(default) the update will be run as regular user.

This solves the problem where running `flatpak update` on a remote
system fails if run as regular user.

Fixes #737.